Add files via upload
The file "Ads_CTR_Optimisation.csv" contains data related to the performance of 10 different advertisements (Ad 1 to Ad 10). Each row represents a user interaction, where a value of 1 indicates that the user clicked on the corresponding ad, and 0 means the user did not click. The dataset is used for optimizing Click-Through Rates (CTR) by analyzing which ads are more likely to be clicked by users. This type of data is commonly used in online advertising to improve ad placement and targeting strategies